requirements.txt
# requirements.txt

# Core Framework
Flask==3.1.0
Werkzeug==3.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
mysql-connector-python==8.3.0
PyMySQL==1.1.0

# Forms and Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0.post1

# Security
python-dotenv==1.0.0
bcrypt==4.1.2

# Date/Time Handling
python-dateutil==2.8.2

# Development Tools
pytest==7.4.4
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
PyMySQL
flask_migrate